✨feat: fetch DTO using JPQL query in repository

+++ src/main/java/com/project/likelion13thbe/domain/product/converter/ProductConverter.java
@@ -30,7 +30,7 @@ public class ProductConverter {
 
     }
 
-    public static ProductResDTO.ProductPreviewResDTO toProductPreviewResDTO(Product product, Double ratingAvg, Integer reviewCount) {
+    public static ProductResDTO.ProductPreviewResDTO toProductPreviewResDTO(Product product, Double ratingAvg, Long reviewCount) {
 
         return ProductResDTO.ProductPreviewResDTO.builder()
                 .id(product.getId())
