+++ src/main/java/com/project/likelion13thbe/domain/product/dto/response/ProductResDTO.java
@@ -15,7 +15,7 @@ public class ProductResDTO {
             Integer price,
             Integer quantity,
             Double ratingAvg,
-            Integer reviewCount,
+            Long reviewCount,
             LocalDateTime createdAt
     ){
     }
